upkeep(alloy-op-evm): move the lib test module to tests.rs
lib.rs was approaching a thousand lines with the test module inline; follow
the block module's mod tests; convention.
